Refactor updateBillableWeight templates
