+++ src/constants/MoveHistory/EventTemplates/UpdateBillableWeight/updateBillableWeight.jsx
@@ -1,12 +1,14 @@
+import React from 'react';
+
 import o from 'constants/MoveHistory/UIDisplay/Operations';
-import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import a from 'constants/MoveHistory/Database/Actions';
 import t from 'constants/MoveHistory/Database/Tables';
+import LabeledDetails from 'pages/Office/MoveHistory/LabeledDetails';
 
 export default {
   action: a.UPDATE,
-  eventName: o.updateBillableWeightAsTIO,
+  eventName: o.updateBillableWeight,
   tableName: t.entitlements,
-  detailsType: d.LABELED,
   getEventNameDisplay: () => 'Updated move',
+  getDetails: (historyRecord) => <LabeledDetails historyRecord={historyRecord} />,
 };
